v2.0-rc.7

+++ module-article/module-article-model/src/main/java/io/jpress/module/article/model/Article.java
@@ -209,8 +209,8 @@ public class Article extends BaseArticle<Article> {
     }
 
     @Override
-    public Long getOrderNumber() {
-        Long order = super.getOrderNumber();
+    public Integer getOrderNumber() {
+        Integer order = super.getOrderNumber();
         return order == null ? 0 : order;
     }
 }
